- changes NewsDto mapping - changes the method of checking the availability of the Internet
